use uint8_t for mrb_bool instead of unsigned int; close #1222

+++ include/mruby/value.h
@@ -7,7 +7,7 @@
 #ifndef MRUBY_VALUE_H
 #define MRUBY_VALUE_H
 
-typedef unsigned int mrb_bool;
+typedef uint8_t mrb_bool;
 
 #ifndef MRB_NAN_BOXING
